test: rename test file
